Rearrangement of staff_role
